+++ docs/src/pages/[platform]/connected-components/storage/fileuploader/examples/BucketFriendly.tsx
@@ -1,0 +1,12 @@
+import { FileUploader } from '@aws-amplify/ui-react-storage';
+
+export const App = () => {
+  return (
+    <FileUploader
+      acceptedFileTypes={['image/*']}
+      bucket="my-bucket"
+      path="public/"
+      maxFileCount={1}
+    />
+  );
+};