+++ packages/e2e/cypress/integration/common/shared.ts
@@ -66,6 +66,14 @@ Given("I'm running the docs page", () => {
   cy.visit('/');
 });
 
+Given('I intercept requests to host including {string}', (host: string) => {
+  cy.intercept({ url: '**' }, (req) => {
+    if (req.headers.host?.includes(host)) {
+      req.alias = host;
+    }
+  });
+});
+
 Given(
   'I intercept {string} with fixture {string}',
   (json: string, fixture: string) => {
@@ -112,6 +120,15 @@ Then(
     cy.wait(`@${method}_REQUEST`)
       .its('response.statusCode')
       .should('eq', +statusCode);
+  }
+);
+
+Then(
+  'I confirm the {string} request was made to host containing {string}',
+  (request: string, hostValue: string) => {
+    cy.wait(`@${request}`).then((interception) => {
+      expect(interception.request.headers.host).to.include(hostValue);
+    });
   }
 );